# Vai encoding benchmark and refactoring