Relates to rancher/rancher#51173

## Changes in this PR

This PR can be better reviewed [commit by commit](https://github.com/rancher/steve/pull/742/commits).

- Abstracts `toBytes` and `fromBytes` into an `Encoding` interface (but still not replaced, that will come in a follow-up PR).
- Adds different encodings, in addition to `gob`.
- Adds a unit tests for consistency and a small benchmark.
- Iterate on the result of the benchmark to improve the performance of the `gob` encoding, as this library has some not very obvious caveats:
   - It's primary focus is inter-process communication, so it's optimized for sending multiple messages over the same link.
   - Every `gob` [Encoder](https://cs.opensource.google/go/go/+/refs/tags/go1.24.5:src/encoding/gob/encoder.go;l=17-25) and [Decoder](https://cs.opensource.google/go/go/+/refs/tags/go1.24.5:src/encoding/gob/decoder.go;l=28-40) hold some complex internal state (unlike JSON) that will make them act differently on successive executions.
   - For this reason, using a one-time Decoder or Encoder is more expensive when using `gob`.
 - Based on the above findings, I applied some optimizations to make the `gobEncoding` produce minimal payloads in a more performant basis.

### Benchmarking

The results correspond to the pure serialization/deserialization. It's not affected by disk performance because all output is discarded, just counting the number of bytes written.

I used 2 types of test data, a sample Pod, and a ~10KB ConfigMap, in order to try to simulate how different object kinds would behave.

#### Before the optimizations:
* Payload: Pod, operation: encoding
```
BenchmarkEncodings/json-pod/encoding-12                   	  561109	      1969 ns/op	       320.0 bytes
BenchmarkEncodings/json+gzip-pod/encoding-12              	   33721	     35211 ns/op	       195.0 bytes
BenchmarkEncodings/gob-pod/encoding-12                    	    7814	    142722 ns/op	     15065 bytes
BenchmarkEncodings/gob+gzip-pod/encoding-12               	    2628	    448266 ns/op	      5696 bytes
```
* Payload: Pod, operation: decoding
```
BenchmarkEncodings/json-pod/decoding-12                   	  376699	      3207 ns/op
BenchmarkEncodings/json+gzip-pod/decoding-12              	  182582	      6552 ns/op
BenchmarkEncodings/gob-pod/decoding-12                    	    4860	    244194 ns/op
BenchmarkEncodings/gob+gzip-pod/decoding-12               	    3778	    320915 ns/op
```
* Payload: ConfigMap, operation: encoding
```
BenchmarkEncodings/json-10KB-configmap/encoding-12        	  119346	      9857 ns/op	     10095 bytes
BenchmarkEncodings/gob-10KB-configmap/encoding-12         	   97942	     11386 ns/op	     10929 bytes
BenchmarkEncodings/json+gzip-10KB-configmap/encoding-12   	    8539	    137156 ns/op	      7340 bytes
BenchmarkEncodings/gob+gzip-10KB-configmap/encoding-12    	    7658	    158344 ns/op	      7945 bytes
```
* Payload: ConfigMap, operation: decoding
```
BenchmarkEncodings/gob-10KB-configmap/decoding-12         	   39224	     30719 ns/op
BenchmarkEncodings/json-10KB-configmap/decoding-12        	   26343	     47619 ns/op
BenchmarkEncodings/gob+gzip-10KB-configmap/decoding-12    	   10000	    109181 ns/op
BenchmarkEncodings/json+gzip-10KB-configmap/decoding-12   	   10000	    117103 ns/op
```
Both performance and payload size is considerably worse using `gob` than JSON (even the gzipped version).
<details>
<summary>Raw output</summary>

```
❯ go test -bench=. ./pkg/sqlcache/db/ -run='BenchmarkEncodings'
goos: darwin
goarch: arm64
pkg: github.com/rancher/steve/pkg/sqlcache/db
cpu: Apple M2 Pro
BenchmarkEncodings/gob-10KB-configmap/encoding-12         	   97942	     11386 ns/op	     10929 bytes
BenchmarkEncodings/gob-10KB-configmap/decoding-12         	   39224	     30719 ns/op
BenchmarkEncodings/gob-pod/encoding-12                    	    7814	    142722 ns/op	     15065 bytes
BenchmarkEncodings/gob-pod/decoding-12                    	    4860	    244194 ns/op
BenchmarkEncodings/json-10KB-configmap/encoding-12        	  119346	      9857 ns/op	     10095 bytes
BenchmarkEncodings/json-10KB-configmap/decoding-12        	   26343	     47619 ns/op
BenchmarkEncodings/json-pod/encoding-12                   	  561109	      1969 ns/op	       320.0 bytes
BenchmarkEncodings/json-pod/decoding-12                   	  376699	      3207 ns/op
BenchmarkEncodings/gob+gzip-10KB-configmap/encoding-12    	    7658	    158344 ns/op	      7945 bytes
BenchmarkEncodings/gob+gzip-10KB-configmap/decoding-12    	   10000	    109181 ns/op
BenchmarkEncodings/gob+gzip-pod/encoding-12               	    2628	    448266 ns/op	      5696 bytes
BenchmarkEncodings/gob+gzip-pod/decoding-12               	    3778	    320915 ns/op
BenchmarkEncodings/json+gzip-10KB-configmap/encoding-12   	    8539	    137156 ns/op	      7340 bytes
BenchmarkEncodings/json+gzip-10KB-configmap/decoding-12   	   10000	    117103 ns/op
BenchmarkEncodings/json+gzip-pod/encoding-12              	   33721	     35211 ns/op	       195.0 bytes
BenchmarkEncodings/json+gzip-pod/decoding-12              	  182582	      6552 ns/op
PASS
ok  	github.com/rancher/steve/pkg/sqlcache/db	19.528s
```

</details>

#### After the optimizations:

* Payload: Pod, operation: encoding
```
BenchmarkEncodings/gob-pod/encoding-12                    	  657393	      1639 ns/op	       109.0 bytes
BenchmarkEncodings/json-pod/encoding-12                   	  595126	      1966 ns/op	       320.0 bytes
BenchmarkEncodings/gob+gzip-pod/encoding-12               	   37135	     32140 ns/op	       103.0 bytes
BenchmarkEncodings/json+gzip-pod/encoding-12              	   33463	     35351 ns/op	       195.0 bytes
```
* Payload: Pod, operation: decoding
```
BenchmarkEncodings/gob-pod/decoding-12                    	 1347354	       889.9 ns/op
BenchmarkEncodings/json-pod/decoding-12                   	  405661	      3030 ns/op
BenchmarkEncodings/gob+gzip-pod/decoding-12               	  353143	      3237 ns/op
ls -lartBenchmarkEncodings/json+gzip-pod/decoding-12              	  188190	      6456 ns/op
```
* Payload: ConfigMap, operation: encoding
```
BenchmarkEncodings/gob-10KB-configmap/encoding-12         	 1479949	       811.0 ns/op	     10035 bytes
BenchmarkEncodings/json-10KB-configmap/encoding-12        	  121186	      9885 ns/op	     10095 bytes
BenchmarkEncodings/gob+gzip-10KB-configmap/encoding-12    	    9298	    126947 ns/op	      7305 bytes
BenchmarkEncodings/json+gzip-10KB-configmap/encoding-12   	    8536	    136483 ns/op	      7340 bytes
```
* Payload: ConfigMap, operation: decoding
```
BenchmarkEncodings/gob-10KB-configmap/decoding-12         	  393547	      2993 ns/op
BenchmarkEncodings/json-10KB-configmap/decoding-12        	   26888	     44560 ns/op
BenchmarkEncodings/gob+gzip-10KB-configmap/decoding-12    	   15278	     77459 ns/op
BenchmarkEncodings/json+gzip-10KB-configmap/decoding-12   	   10000	    116002 ns/op
```

<details>
<summary>Raw output</summary>

```
❯ go test -bench=. ./pkg/sqlcache/db/ -run='BenchmarkEncodings'
goos: darwin
goarch: arm64
pkg: github.com/rancher/steve/pkg/sqlcache/db
cpu: Apple M2 Pro
BenchmarkEncodings/gob-10KB-configmap/encoding-12         	 1479949	       811.0 ns/op	     10035 bytes
BenchmarkEncodings/gob-10KB-configmap/decoding-12         	  393547	      2993 ns/op
BenchmarkEncodings/gob-pod/encoding-12                    	  657393	      1639 ns/op	       109.0 bytes
BenchmarkEncodings/gob-pod/decoding-12                    	 1347354	       889.9 ns/op
BenchmarkEncodings/json-10KB-configmap/encoding-12        	  121186	      9885 ns/op	     10095 bytes
BenchmarkEncodings/json-10KB-configmap/decoding-12        	   26888	     44560 ns/op
BenchmarkEncodings/json-pod/encoding-12                   	  595126	      1966 ns/op	       320.0 bytes
BenchmarkEncodings/json-pod/decoding-12                   	  405661	      3030 ns/op
BenchmarkEncodings/gob+gzip-10KB-configmap/encoding-12    	    9298	    126947 ns/op	      7305 bytes
BenchmarkEncodings/gob+gzip-10KB-configmap/decoding-12    	   15278	     77459 ns/op
BenchmarkEncodings/gob+gzip-pod/encoding-12               	   37135	     32140 ns/op	       103.0 bytes
BenchmarkEncodings/gob+gzip-pod/decoding-12               	  353143	      3237 ns/op
BenchmarkEncodings/json+gzip-10KB-configmap/encoding-12   	    8536	    136483 ns/op	      7340 bytes
BenchmarkEncodings/json+gzip-10KB-configmap/decoding-12   	   10000	    116002 ns/op
BenchmarkEncodings/json+gzip-pod/encoding-12              	   33463	     35351 ns/op	       195.0 bytes
ls -lartBenchmarkEncodings/json+gzip-pod/decoding-12              	  188190	      6456 ns/op
PASS
ok  	github.com/rancher/steve/pkg/sqlcache/db	19.334s
```

</details>

#### Conclusions
1. Payload size (only considering small objects; bigger payloads would make any improvement negligible).
  - vs. one-time gob encoders: **100x** smaller
  - vs. JSON: **3x smaller**.
3. Encoding:
  - vs. one-time gob encoders: **~87x faster** for small objects, 14x faster for 10KB ConfigMaps.
  - vs. JSON: 1.2x faster for small objects, **12x** faster for 10KB ConfigMaps.
4. Decoding:
  - vs. one-time gob encoders: **~275x faster** for small objects, ~10x faster for 10KB ConfigMaps.
  - vs. JSON: 3.4x faster for small objects, **~15x** faster for 10KB ConfigMaps.